+++ HomePage.h
@@ -1,0 +1,19 @@
+#pragma once
+#include <iostream>
+#include "Librarian.h"
+
+class HomePage
+{
+
+public:
+	HomePage() {
+
+		std::cout << "1.Client     2.Librarian\n";
+
+
+	}
+
+
+
+};
+
