+++ LIBRARY SYSTEM.vcxproj
@@ -24,6 +24,7 @@
     <ProjectGuid>{f76c9c52-db06-4b63-a914-c8264e53433a}</ProjectGuid>
     <RootNamespace>LIBRARYSYSTEM</RootNamespace>
     <WindowsTargetPlatformVersion>10.0</WindowsTargetPlatformVersion>
+    <ProjectName>Library System</ProjectName>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'" Label="Configuration">
@@ -127,7 +128,12 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="LIBRARY SYSTEM.cpp" />
+    <ClCompile Include="Library System.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Client.h" />
+    <ClInclude Include="HomePage.h" />
+    <ClInclude Include="Librarian.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
