+++ Librarian.h
@@ -1,0 +1,19 @@
+#pragma once
+#include "Client.h"
+
+
+class Librarian : public Client
+{
+private:
+
+
+
+public:
+	Librarian() {
+
+
+
+	}
+
+};
+
